+++ directdm/__init__.py
@@ -1,3 +1,4 @@
 from . import run
 from . import num
-from .wilson_coefficients import WC_3f, WC_4f, WC_5f
+from . import match
+from .wilson_coefficients import WC_3f, WC_4f, WC_5f, WC_EW