+++ directdm/run/rge.py
@@ -64,106 +64,96 @@ class QCD_gamma(object):
 class AlphaS(object):
     """ The strong coupling constant """
 
-    def __init__(self, nf, loop):
-        self.nf = nf
-        self.loop = loop
+    def __init__(self, asMZ, MZ):
+        self.asMZ = asMZ
+        self.MZ = MZ
 
-        #----------------------#
-        #--- Some constants ---#
-        #----------------------#
-
-        ip = Num_input()
-        self.MZ = ip.Mz
-        self.asMZ = ip.asMZ
-        self.mc_at_2GeV = ip.mc_at_2GeV #GeV
-        self.mb_at_mb = ip.mb_at_mb #GeV
-
-    def decouple_down_MSbar(self, alphasatmu, mu, mh):
-        """ Decoupling of the strong coupling from nf to (nf - 1) at scale mu, at heavy quark mass mh
+    def decouple_down_MSbar(self, alphasatmu, mu, mh, nf, loop):
+        """ Decoupling of the strong coupling from nf to (nf - 1) at scale mu, at heavy quark mass mh = mh(mh)
 
         Input is alphas(mu,nf), output is alphas(mu,nf-1)
         """
-        if self.loop == 1:
+        if loop == 1:
             return alphasatmu
-        if self.loop == 2:
+        if loop == 2:
             return alphasatmu * (1 - 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi))
-        if self.loop == 3:
+        if loop == 3:
             return alphasatmu * (1 - 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
-                                   + (11/72 - 11/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2)
-        if self.loop == 4:
+                                   + (11/72 - 19/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2)
+        if loop == 4:
             return alphasatmu * (1 - 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
-                                   + (11/72 - 11/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2
-                                   + (564731/124416 - 82043/27648 * my_zeta(3) - 955/576 * np.log(mu**2/mh**2) + 53/576 * np.log(mu**2/mh**2)**2
-                                      - 1/216 * np.log(mu**2/mh**2)**3 + (self.nf-1) * ( -2633/31104 + 67/576 * np.log(mu**2/mh**2)
-                                      - 1/36 * np.log(mu**2/mh**2)**2 ) ) * (alphasatmu/np.pi)**3 )
+                                   + (11/72 - 19/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2
+                                   + (564731/124416 - 82043/27648 * my_zeta(3) - 6793/1728 * np.log(mu**2/mh**2) - 131/576 * np.log(mu**2/mh**2)**2
+                                      - 1/216 * np.log(mu**2/mh**2)**3 + (nf-1) * ( -2633/31104 + 281/1728 * np.log(mu**2/mh**2)
+                                      ) ) * (alphasatmu/np.pi)**3 )
 
+    # def decouple_up_MSbar(self, alphasatmu, mu, mh):
+    #     """ Decoupling of the strong coupling from (nf-1) to nf at scale mu, at heavy quark mass mh = mh(mh)
 
-    def decouple_up_MSbar(self, alphasatmu, mu, mh):
-        """ Decoupling of the strong coupling from (nf-1) to nf at scale mu, at heavy quark mass mh
+    #     Input is alphas(mu,nf-1), output is alphas(mu,nf)
+    #     """
+    #     if loop == 1:
+    #         return alphasatmu
+    #     if loop == 2:
+    #         return alphasatmu * (1 + 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi))
+    #     if loop == 3:
+    #         return alphasatmu * (1 + 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
+    #                                + (- 11/72 + 19/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2)
+    #     if loop == 4:
+    #         return alphasatmu * (1 - 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
+    #                                + (- 11/72 + 19/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2
+    #                                + (- 564731/124416 + 82043/27648 * my_zeta(3) + 2191/576 * np.log(mu**2/mh**2) + 511/576 * np.log(mu**2/mh**2)**2
+    #                                   + 1/216 * np.log(mu**2/mh**2)**3 + (self.nf-1) * ( 2633/31104 - 281/1728 * np.log(mu**2/mh**2)
+    #                                   ) ) * (alphasatmu/np.pi)**3 )
 
-        Input is alphas(mu,nf-1), output is alphas(mu,nf)
-        """
-        if self.loop == 1:
-            return alphasatmu
-        if self.loop == 2:
-            return alphasatmu * (1 + 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi))
-        if self.loop == 3:
-            return alphasatmu * (1 + 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
-                                   + (- 11/72 + 11/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2)
-        if self.loop == 4:
-            return alphasatmu * (1 - 1/6 * np.log(mu**2/mh**2) * (alphasatmu/np.pi)
-                                   + (- 11/72 + 11/24 * np.log(mu**2/mh**2) + 1/36 * np.log(mu**2/mh**2)**2) * (alphasatmu/np.pi)**2
-                                   + (- 564731/124416 + 82043/27648 * my_zeta(3) + 2645/1728 * np.log(mu**2/mh**2) + 167/576 * np.log(mu**2/mh**2)**2
-                                      + 1/216 * np.log(mu**2/mh**2)**3 + (self.nf-1) * ( 2633/31104 - 67/576 * np.log(mu**2/mh**2)
-                                      + 1/36 * np.log(mu**2/mh**2)**2 ) ) * (alphasatmu/np.pi)**3 )
-
-    def __dalphasdmu(self, mu, alphas, nf):
-        if self.loop == 1:
+    def __dalphasdmu(self, mu, alphas, nf, loop):
+        if loop == 1:
             return 2 * np.pi / mu * ( - QCD_beta(nf, 1).chet() * (alphas/np.pi)**2 )
-        if self.loop == 2:
+        if loop == 2:
             return 2 * np.pi / mu * ( - QCD_beta(nf, 1).chet() * (alphas/np.pi)**2 - QCD_beta(nf, 2).chet() * (alphas/np.pi)**3 )
-        if self.loop == 3:
-            return 2 * np.pi / mu * ( - QCD_beta(nf, 1).chet() * (alphas/np.pi)**2 - QCD_beta(nf, 2).chet() * (alphas/np.pi)**3 - QCD_beta(nf, 3).chet() * (alphas/np.pi)**4 )
+        if loop == 3:
+            return 2 * np.pi / mu * ( - QCD_beta(nf, 1).chet() * (alphas/np.pi)**2 - QCD_beta(nf, 2).chet() * (alphas/np.pi)**3\
+                                      - QCD_beta(nf, 3).chet() * (alphas/np.pi)**4 )
 
-    def solve_rge(self, as_at_mu, mu, mu0):
+    def __solve_rge_nf(self, as_at_mu, mu, mu0, nf, loop):
         """The running strong coupling
     
-        Run from scale mu to scale mu0, initial condition alphas(mu) = as_at_mu
+        Run from scale mu to scale mu0, with nf active flavors
         """
         def deriv(alphas, mu):
-            return self.__dalphasdmu(mu, alphas, self.nf)
+            return self.__dalphasdmu(mu, alphas, nf, loop)
         r = odeint(deriv, as_at_mu, np.array([mu, mu0]))
         return list(r)[1][0]
 
-    def __solve_rge_nf(self, as_at_mu, mu, mu0, nf):
-        """The running strong coupling
-    
-        Run from scale mu to scale mu0, with initial condition alphas(mu) = as_at_mu, and nf active flavors
+    def run(self, dict_mh, dict_mu, mu0, nf, loop):
+        """ Run the strong coupling with decoupling at flavor thresholds
+
+        A dictionary of scales for each heavy quark mass mq(mq) should be given. E.g.
+
+        {'mbmb': 4.18, 'mcmc': 1.275}
+
+        A dictionary of scales for each threshold should be given. E.g.
+
+        {'mub': 5, 'muc': 1.3}
+
+        (Depending on nf one or zero can be given)
+
+        The decoupling is always at mq(mq)
         """
-        def deriv(alphas, mu):
-            return self.__dalphasdmu(mu, alphas, nf)
-        r = odeint(deriv, as_at_mu, np.array([mu, mu0]))
-        return list(r)[1][0]
+        if nf == 5:
+            return self.__solve_rge_nf(self.asMZ, self.MZ, mu0, 5, loop)
+        if nf == 4:
+            as5_mub = self.__solve_rge_nf(self.asMZ, self.MZ, dict_mu['mub'], 5, loop)
+            as4_mub = self.decouple_down_MSbar(as5_mub, dict_mu['mub'], dict_mh['mbmb'], 5, loop)
+            return self.__solve_rge_nf(as4_mub, dict_mu['mub'], mu0, 4, loop)
+        if nf == 3:
+            as5_mub = self.__solve_rge_nf(self.asMZ, self.MZ, dict_mu['mub'], 5, loop)
+            as4_mub = self.decouple_down_MSbar(as5_mub, dict_mu['mub'], dict_mh['mbmb'], 5, loop)
+            as4_muc = self.__solve_rge_nf(as4_mub, dict_mu['mub'], dict_mu['muc'], 4, loop)
+            as3_muc = self.decouple_down_MSbar(as4_muc, dict_mu['muc'], dict_mh['mcmc'], 4, loop)
+            return self.__solve_rge_nf(as3_muc, dict_mu['muc'], mu0, 3, loop)
 
-    def run(self, mu0):
-        """ Run the strong coupling 
-
-        Start value is alphas(MZ), decoupling at quark thresholds is performed automatically. 
-        """
-        if self.nf == 5:
-            return self.solve_rge(self.asMZ, self.MZ, mu0)
-        if self.nf == 4:
-            as5_mb = self.__solve_rge_nf(self.asMZ, self.MZ, self.mb_at_mb, 5)
-            as4_mb = self.decouple_down_MSbar(as5_mb, self.mb_at_mb, self.mb_at_mb)
-            return self.__solve_rge_nf(as4_mb, self.mb_at_mb, mu0, 4)
-        if self.nf == 3:
-            as5_mb = self.__solve_rge_nf(self.asMZ, self.MZ, self.mb_at_mb, 5)
-            as4_mb = self.decouple_down_MSbar(as5_mb, self.mb_at_mb, self.mb_at_mb)
-            as4_mc = self.__solve_rge_nf(as4_mb, self.mb_at_mb, 2, 4)
-            as3_mc = self.decouple_down_MSbar(as4_mc, 2, self.mc_at_2GeV)
-            return self.__solve_rge_nf(as4_mc, 2, mu0, 3)
-
-### Future: class should be C_QCD, given the Wilson coefficient at different scales
+### Future: class should be CmuQCD, giving the Wilson coefficient at different scales
 
 class RGE(object):
 
@@ -181,4 +171,127 @@ class RGE(object):
         b0 = QCD_beta(self.nf, 1).trad()
         return expm((asmuh-asmul)/(4*np.pi) * np.array(np.transpose(self.adm[0]))/(2*b0))
 
+
+
+class CmuEW(object):
+    def __init__(self, Wilson, ADM, muh, mul, Y, d):
+        """ Calculate the running of the Wilson coefficients in the unbroken EW theory
+
+        The running takes into account the gauge coupling g1, g2, g3,
+        the tau, bottom, and top Yukawas, ytau, yb, yt, and the Higgs self coupling lambda.
+
+        Wilson should be a list of initial conditions for the Wilson coefficients.
+
+        ADM should be a list / array of the seven DM anomalous dimension matrices proportional to g1^2, g2^2, g3^2, ytau^2, yb^2, yt^2, lambda.
+
+        muh is the initial scale.
+
+        mul is the final scale.
+        """
+        self. Wilson = Wilson
+        self.ADM = ADM
+        self.muh = muh
+        self.mul = mul
+        self.Y = Y
+        self.d = d
+
+        # Input parameters
+
+        ip = Num_input()
+
+        self.g1   = ip.g1_at_MZ
+        self.g2   = ip.g2_at_MZ
+        self.gs   = ip.g3_at_MZ
+        self.ytau = ip.ytau_at_MZ
+        self.yc   = ip.yc_at_MZ
+        self.yb   = ip.yb_at_MZ
+        self.yt   = ip.yt_at_MZ
+        self.lam  = ip.lam_at_MZ
+
+        # The initial values of the couplings at MZ
+        self.MZ    = ip.Mz
+        self.ginit = [self.g1, self.g2, self.gs, self.yc, self.ytau, self.yb, self.yt, self.lam]
+
+
+    def _dgdmu(self, g, mu, Y, d):
+        """ Calculate the log derivative [i.e. dg/dlog(mu)] of the couplings g1, g2, g3, yc, ytau, yb, yt, lam w.r.t. to mu, at scale mu
+        
+        Take a 8-vector (list) of couplings g = [g1, g2, g3, yc, ytau, yb, yt, lambda]
+
+        Take the DM quantum numbers d, Y (so far only 1 multiplet)
+
+        Return the derivative -- again a 8-vector
+        """
+        N = 1
+        # The 8x8 matrix of beta functions (Arason et al., Phys.Rev. D46 (1992) 3945-3965, and our calculation)
+        # Note the different sign and normalization conventions. (g1_Arason = 5/3 * g1_Denner; lambda_Arason = 1/4 * lambda_Denner)
+
+        # g1, g2, g3, yc, ytau, yb, yt
+        g7 = np.array(g[:-1])
+        g7_squared = np.array(list(map(lambda x: x**2, g[:-1])))
+        beta = np.array([[41/6+Y**2*d*N/3, 0,                         0,  0,   0,    0,    0  ],
+                         [0,               -19/6+4*(d**2-1)/4*d*N/9,  0,  0,   0,    0,    0  ],
+                         [0,               0,                        -7, 0,   0,    0,    0  ],
+                         [-17/12,          -9/4,                     -8, 9/2, 1,    3/2,  9/2],
+                         [-15/4,           -9/4,                      0,  3,   5/2,  3,    3  ],
+                         [-5/12,           -9/4,                     -8, 3/2, 1,    9/2,  3/2],
+                         [-17/12,          -9/4,                     -8, 9/2, 1,    3/2,  9/2]])
+
+        # g1, g2, g3, yc, ytau, yb, yt, lambda
+        beta_lam_1 = np.array([-3/4, -9/4, 0, 3, 1, 3, 3, 3/4])
+
+        # g1^2, g2^2, g3^2, yc^2, ytau^2, yb^2, yt^2
+        beta_lam_2 = np.array([[3/4,  3/4,  0,  0,   0,  0,   0  ],
+                               [3/4,  9/4,  0,  0,   0,  0,   0  ],
+                               [0,    0,    0,  0,   0,  0,   0  ],
+                               [0,    0,    0,  -12, 0,  0,   0  ],
+                               [0,    0,    0,  0,   -4, 0,   0  ],
+                               [0,    0,    0,  0,   0,  -12, 0  ],
+                               [0,    0,    0,  0,   0,  0,   -12]])
+
+        deriv_list = np.hstack(( np.multiply(g7, np.dot(beta, g7_squared)),\
+                                 np.multiply(g[7], np.dot(beta_lam_1[:-1], g7_squared))\
+                                 + beta_lam_1[7]*g[7]**2 + np.dot(g7_squared, np.dot(beta_lam_2, g7_squared)) )) / mu / (4*np.pi)**2
+
+        return deriv_list
+
+    def _alphai(self, g_init, mu_init, mu2, Y, d):
+        """ Calculate the one-loop running of alpha1, alpha2, alpha3, alphac, alphatau, alphab, alphat, lambda in the six-flavor theory
+        
+        Run from mu_init to mu2. ginit are the couplings defined at scale mu_init.
+
+        Take g's as input and return alpha's
+        """
+
+
+        # Runge-Kutta:        
+        def deriv(mu, g):
+            return self._dgdmu(g, mu, Y, d)
+        g0, mu0 = g_init, mu_init
+        dmu = mu2 - mu_init
+#        r = ode(deriv).set_integrator('lsoda')
+        r = ode(deriv).set_integrator('dopri5')
+#        r = ode(deriv).set_integrator('vode')
+#        r = ode(deriv).set_integrator('dop853')
+        r.set_initial_value(g0, mu0)
+        solution = r.integrate(r.t+dmu)
+        alpha = np.hstack((np.array(list(map(lambda x: x**2/4/np.pi, solution[:-1]))), np.array([solution[-1]/4/np.pi])))
+
+        return alpha
+
+
+    def run(self):
+
+        def deriv(mu, C):
+            return sum([np.dot(C,self.ADM[k])*self._alphai(self.ginit, self.MZ, mu, self.Y, self.d)[k]/4/np.pi/mu for k in range(8)])
+
+        C0, mu0 = self.Wilson, self.muh
+        dmu = self.mul - self.muh
+        r = ode(deriv).set_integrator('dopri5')
+#        r = ode(deriv).set_integrator('lsoda')
+#        r = ode(deriv).set_integrator('dop853')
+#        r = ode(deriv).set_integrator('vode')
+        r.set_initial_value(C0, mu0)
+        C = np.array(r.integrate(r.t+dmu))
+        return(C)